requirements.txt
# Core Data Processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Data Acquisition
kagglehub==0.2.3
yfinance==0.2.48

# Technical Analysis
ta==0.11.0
scipy==1.11.1
scikit-learn==1.3.0

# Database
sqlalchemy==2.0.19
alembic==1.11.1

# API Layer
fastapi==0.101.0
uvicorn[standard]==0.23.2
pydantic==2.1.1
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.0
tqdm==4.66.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Development Dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.4.1
ipython==8.14.0
jupyter==1.0.0
